requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bootstrap4==0.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==5.1.1
Django==4.1.1
django-bootstrap4==22.2
django-bootstrap5==22.1
django-environ==0.9.0
idna==3.4
numpy==1.23.3
Pillow==9.2.0
pyshorteners==1.0.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
validators==0.20.0
environ
opencv-python-headless
backports.zoneinfo;python_version<"3.9"
gunicorn
psycopg2
python-dotenv
whitenoise